requirements.txt
streamlit==1.41.1
SQLAlchemy==2.0.36
pandas==2.2.3
xlsxwriter==3.2.0
Pillow==10.4.0

bcrypt==4.2.0
easyocr==1.7.2

google-generativeai==0.8.3
